pothos-blocks: network special handling for packet type
Because the network blocks serialize messages, messages are not as efficient as stream buffers.
When we switched to portable serialization this didnt get any better.
And since the packet type is so widely used with ordinary buffers which are POD,
the network blocks now split packet messages into a serialized header and POD data.